Move retrieval stream converter to implementation
Keep knowledge base stream event dispatch and payload decoding in the unexported knowledgebases implementation package while preserving generated client wiring.


Copilot-Session: eecc64df-f690-4410-b68c-b45e323ec918
